refactor: 에러 코드 및 경로 개선
