+++ src/main/java/com/market/saessag/domain/user/entity/User.java
@@ -53,4 +53,9 @@ public class User {
         this.updatedAt = LocalDateTime.now();
     }
 
+    public void updatePassword(String newPassword) {
+        this.password = newPassword;
+        this.updatedAt = LocalDateTime.now();
+    }
+
 }
